Remove event mangling
Export decoding functions for warp event encoding/decoding
