+++ tests/behaviour/expectations/types.ts
@@ -2,9 +2,9 @@ import assert from 'assert';
 import path from 'path';
 import { mangleContractFilePath } from '../../../src/passes/sourceUnitSplitter';
 import { stringFlatten } from './utils';
+import { EventItem } from '../../../src/utils/event';
 
 export const OUTPUT_DIR = 'warp_output';
-export type EventItem = { data: string[]; keys: string[]; order: number };
 
 export class AsyncTest {
   constructor(
